plugins/cusum/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift-cusum"
version = "2.0.0"
description = "CUSUM drift detection plugin for drift-framework"
requires-python = ">=3.11"
dependencies = [
    "drift-framework>=0.1,<1.0",
    "matplotlib>=3.8",
]

[project.entry-points."drift_framework.plugins"]
cusum = "drift_cusum:register"

[tool.setuptools.packages.find]
include = ["drift_cusum*"]

[tool.setuptools.package-data]
drift_cusum = [
    "web/templates/**/*",
    "web/static/**/*",
    "examples/*.csv",
]
